requirements.txt
Flask==0.10.1
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauthlib==0.7.2
python-dateutil==2.4.2
requests==2.6.2
requests-oauthlib==0.4.2
six==1.9.0
Werkzeug==0.10.4
